+++ components/utils/LoginDialog.vue
@@ -4,7 +4,7 @@
       <v-card>
         <v-card-title>Sign in</v-card-title>
         <v-divider class="my-2 py-2"></v-divider>
-        <v-card-text>
+        <v-card-text class="my-0 py-0">
           <form method="post" @submit.prevent="login">
             <v-layout column>
               <v-flex>
@@ -26,9 +26,6 @@
                   type="password"
                   required
                 ></v-text-field>
-              </v-flex>
-              <v-flex class="text-xs-center" my-5>
-                <v-btn color="primary" type="submit"> Submit </v-btn>
               </v-flex>
               <p>
                 Don't have an account?
@@ -52,7 +49,7 @@
         </v-card-text>
 
         <v-card-actions>
-          <v-btn class="mx-2" color="primary" dark @click="login()">
+          <v-btn class="mx-2 mb-2" color="primary" dark @click="login()">
             sign in </v-btn
           ><v-btn color="grey darken-1" text @click="close()"> cancel </v-btn>
           <v-spacer></v-spacer>
@@ -81,10 +78,45 @@ export default {
     }
   },
   computed: {},
+  watch: {
+    dialog(val) {
+      !val && this.$emit('close')
+    },
+  },
   methods: {
-    login() {
-      this.$emit('close')
-      return 1
+    async login() {
+      try {
+        await this.$auth.loginWith('local', {
+          data: {
+            username: this.email,
+            password: this.password,
+          },
+        })
+        this.$router.push('/')
+        this.$store.commit('SET_TOKEN', this.$auth.getToken('local'))
+        this.$axios.setToken(this.$auth.getToken('local'))
+        this.$emit('close')
+        console.log(
+          this.$store.getters.loggedInUser,
+          this.$auth.getToken('local')
+        )
+        this.$store.dispatch(
+          'fetchProfile',
+          this.$store.getters.loggedInUser,
+          this.$auth.getToken('local')
+        )
+        this.$notifier.showMessage({
+          content: 'Sign in success!',
+          color: 'success',
+        })
+      } catch (e) {
+        this.statusCode = e.response.data.statusCode
+        this.error = e.response.data.message
+        this.$notifier.showMessage({
+          content: 'Login failed!',
+          color: 'error',
+        })
+      }
     },
     close() {
       this.$emit('close')